Initial implementation of authorization flow
This is the first outline of access control in the V2 registry. The main point
of interest is the AccessController, which gates access to resources via http
requests. The sillyAccessController gates requests on whether or not they
simply have the Authorization header set for the request.

Testing is minimal, at this point. This commit will likely be binned, broken
down into a few parts and be fully tested.